Fix and update cli
