+++ pyproject.toml
@@ -50,6 +50,10 @@ sphinx = { version = "^7", optional = true }
 sphinx_rtd_theme = { version = "^1", optional = true }
 sphinxcontrib-programoutput = { version = "^0", optional = true }
 docutils = { version = ">=0.17", optional = true }
+########################
+
+asyncclick = "^8.1.3.4"
+anyio = "^4.0.0" # see https://github.com/python-trio/asyncclick/issues/18
 
 [tool.poetry.group.dev.dependencies]
 flake8 = "*"
