+++ ring_doorbell/cli.py
@@ -1,24 +1,52 @@
 # vim:sw=4:ts=4:et
 # Many thanks to @troopermax <https://github.com/troopermax>
-"""Python Ring CLI."""
+"""Python Ring command line interface."""
 import json
 import getpass
-import argparse
+import asyncio
 from pathlib import Path
-from oauthlib.oauth2 import MissingTokenError
-from ring_doorbell import Ring, Auth
+from oauthlib.oauth2 import MissingTokenError, InvalidGrantError
+import asyncclick as click
+
+from ring_doorbell.auth import Auth
+from ring_doorbell.ring import Ring
+from ring_doorbell.const import USER_AGENT, CLI_TOKEN_FILE
 
 
 def _header():
     _bar()
-    print("Ring CLI")
+    echo("Ring CLI")
 
 
 def _bar():
-    print("---------------------------------")
+    echo("---------------------------------")
 
 
-cache_file = Path("test_token.cache")
+click.anyio_backend = "asyncio"
+
+pass_ring = click.make_pass_decorator(Ring)
+
+
+class ExceptionHandlerGroup(click.Group):
+    """Group to capture all exceptions and echo them nicely.
+
+    Idea from https://stackoverflow.com/a/44347763
+    """
+
+    def __call__(self, *args, **kwargs):
+        """Run the coroutine in the event loop and echo any exceptions."""
+        try:
+            loop = asyncio.new_event_loop()
+            asyncio.set_event_loop(loop)
+            loop.run_until_complete(self.main(*args, **kwargs))
+
+        except Exception as ex:  # pylint: disable=broad-exception-caught
+            echo(f"Got error: {ex!r}")
+
+
+cache_file = Path(CLI_TOKEN_FILE)
+
+echo = click.echo
 
 
 def token_updated(token):
@@ -43,68 +71,134 @@ def _format_filename(event):
     return filename
 
 
-def cli():
-    """Command line function."""
-    parser = argparse.ArgumentParser(
-        description="Ring Doorbell",
-        epilog="https://github.com/tchellomello/python-ring-doorbell",
-        formatter_class=argparse.RawDescriptionHelpFormatter,
-    )
+def _do_auth(username, password):
+    if not username:
+        username = input("Username: ")
 
-    parser.add_argument(
-        "-u", "--username", dest="username", type=str, help="username for Ring account"
-    )
+    if not password:
+        password = getpass.getpass("Password: ")
 
-    parser.add_argument(
-        "-p", "--password", type=str, dest="password", help="username for Ring account"
-    )
+    auth = Auth("RingCLI/0.6", None, token_updated)
+    try:
+        auth.fetch_token(username, password)
+        return auth
+    except MissingTokenError:
+        auth.fetch_token(username, password, input("2FA Code: "))
+        return auth
 
-    parser.add_argument(
-        "--count",
-        action="store_true",
-        default=False,
-        help="count the number of videos on your Ring account",
-    )
 
-    parser.add_argument(
-        "--download-all",
-        action="store_true",
-        default=False,
-        help="download all videos on your Ring account",
-    )
-
-    args = parser.parse_args()
-    _header()
-
+def _get_updated_ring(username, password):
     # connect to Ring account
     if cache_file.is_file():
         auth = Auth(
-            "RingCLI/0.6",
+            USER_AGENT,
             json.loads(cache_file.read_text(encoding="utf-8")),
             token_updated,
         )
-    else:
-        if not args.username:
-            args.username = input("Username: ")
-
-        if not args.password:
-            args.password = getpass.getpass("Password: ")
-
-        auth = Auth("RingCLI/0.6", None, token_updated)
+        ring = Ring(auth)
         try:
-            auth.fetch_token(args.username, args.password)
-        except MissingTokenError:
-            auth.fetch_token(args.username, args.password, input("2FA Code: "))
+            ring.update_data()
+        except InvalidGrantError:
+            auth = _do_auth(username, password)
+            ring = Ring(auth)
+            ring.update_data()
+    else:
+        auth = _do_auth(username, password)
+        ring = Ring(auth)
+        ring.update_data()
 
-    ring = Ring(auth)
-    ring.update_data()
+    return ring
+
+
+@click.group(
+    invoke_without_command=True,
+    cls=ExceptionHandlerGroup,
+)
+@click.option(
+    "--username",
+    default=None,
+    required=False,
+    envvar="RING_USERNAME",
+    help="Username for Ring account.",
+)
+@click.option(
+    "--password",
+    default=None,
+    required=False,
+    envvar="RING_PASSWORD",
+    help="Password for Ring account",
+)
+@click.version_option(package_name="ring_doorbell")
+@click.pass_context
+async def cli(ctx, username, password):
+    """Command line function."""
+
+    _header()
+
+    ring = _get_updated_ring(username, password)
+    ctx.obj = ring
+
+    if ctx.invoked_subcommand is None:
+        return await ctx.invoke(show)
+
+
+@cli.command()
+@pass_ring
+async def show(ring):
+    """Display ring devices."""
+    devices = ring.devices()
+
+    echo(devices)
+
+    doorbells = devices["doorbots"]
+    chimes = devices["chimes"]
+    stickup_cams = devices["stickup_cams"]
+
+    echo(doorbells)
+    echo(chimes)
+    echo(stickup_cams)
+    devices = ring.devices()
+    echo(devices)
+
+    doorbells = devices["doorbots"]
+    chimes = devices["chimes"]
+    stickup_cams = devices["stickup_cams"]
+
+    echo(doorbells)
+    echo(chimes)
+    echo(stickup_cams)
+
+
+@cli.command()
+@click.option(
+    "--count",
+    required=False,
+    default=False,
+    is_flag=True,
+    help="Count the number of videos on your Ring account",
+)
+@click.option(
+    "--download-all",
+    required=False,
+    default=False,
+    is_flag=True,
+    help="Download all videos on your Ring account",
+)
+@click.option(
+    "--max-count",
+    required=False,
+    default=300,
+    help="Maximum count of videos to count or download from your Ring account",
+)
+@pass_ring
+@click.pass_context
+async def videos(ctx, ring, count, download_all, max_count):
+    """Intercat with ring videos."""
     devices = ring.devices()
     doorbell = devices["doorbots"][0]
 
-    _bar()
-
-    if args.count:
-        print(
+    if count:
+        echo(
             "\tCounting videos linked on your Ring account.\n"
             + "\tThis may take some time....\n"
         )
@@ -115,21 +209,23 @@ def cli():
         while len(history) > 0:
             events += history
             counter += len(history)
-            history = doorbell.history(older_than=history[-1]["id"])
+            history = doorbell.history(older_than=history[-1]["id"], limit=100)
+            if len(events) >= max_count:
+                break
 
         motion = len([m["kind"] for m in events if m["kind"] == "motion"])
         ding = len([m["kind"] for m in events if m["kind"] == "ding"])
         on_demand = len([m["kind"] for m in events if m["kind"] == "on_demand"])
 
-        print("\tTotal videos: {}".format(counter))
-        print("\tDing triggered: {}".format(ding))
-        print("\tMotion triggered: {}".format(motion))
-        print("\tOn-Demand triggered: {}".format(on_demand))
+        echo("\tTotal videos: {}".format(counter))
+        echo("\tDing triggered: {}".format(ding))
+        echo("\tMotion triggered: {}".format(motion))
+        echo("\tOn-Demand triggered: {}".format(on_demand))
 
         # already have all events in memory
-        if args.download_all:
+        if download_all:
             counter = 0
-            print(
+            echo(
                 "\tDownloading all videos linked on your Ring account.\n"
                 + "\tThis may take some time....\n"
             )
@@ -137,21 +233,21 @@ def cli():
             for event in events:
                 counter += 1
                 filename = _format_filename(event)
-                print("\t{}/{} Downloading {}".format(counter, len(events), filename))
+                echo("\t{}/{} Downloading {}".format(counter, len(events), filename))
 
                 doorbell.recording_download(
                     event["id"], filename=filename, override=False
                 )
 
-    if args.download_all and not args.count:
-        print(
+    if download_all and not count:
+        echo(
             "\tDownloading all videos linked on your Ring account.\n"
             + "\tThis may take some time....\n"
         )
         history = doorbell.history(limit=100)
 
         while len(history) > 0:
-            print(
+            echo(
                 "\tProcessing and downloading the next "
                 + format(len(history))
                 + " videos"
@@ -161,7 +257,7 @@ def cli():
             for event in history:
                 counter += 1
                 filename = _format_filename(event)
-                print("\t{}/{} Downloading {}".format(counter, len(history), filename))
+                echo("\t{}/{} Downloading {}".format(counter, len(history), filename))
 
                 doorbell.recording_download(
                     event["id"], filename=filename, override=False
@@ -171,4 +267,4 @@ def cli():
 
 
 if __name__ == "__main__":
-    cli()
+    cli()  # pylint: disable=no-value-for-parameter
