+++ ring_doorbell/stickup_cam.py
@@ -3,7 +3,7 @@
 """Python Ring Doorbell wrapper."""
 import logging
 
-from ring_doorbell import RingDoorBell
+from ring_doorbell.doorbot import RingDoorBell
 from ring_doorbell.const import (
     LIGHTS_ENDPOINT,
     MSG_ALLOWED_VALUES,
